+++ weaver/__meta__.py
@@ -1,7 +1,7 @@
 __name__ = "weaver"
 __title__ = "Weaver"
 __package__ = "crim-weaver"
-__version__ = "6.8.1"
+__version__ = "6.9.0"
 __description__ = "Workflow Execution Management Service (EMS); Application, Deployment and Execution Service (ADES)."
 __source_repository__ = "https://github.com/crim-ca/weaver"
 __docker_repository__ = "https://hub.docker.com/r/pavics/weaver"